% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freem.R
\name{calcular_indice_masculinidad}
\alias{calcular_indice_masculinidad}
\title{Calcular el indice de masculinidad}
\usage{
calcular_indice_masculinidad(hembras, machos)
}
\arguments{
\item{hembras}{Numero de hembras}

\item{machos}{Numero de machos}
}
\value{
El indice de masculinidad en el numero de animales.
}
\description{
La función calcula el indice de masculinidad de un ternero.
}
