% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vef_freem.R
\name{es_freemartin}
\alias{es_freemartin}
\title{Verificar si un animal es un freemartin}
\usage{
es_freemartin(indice_masculinidad, umbral = 0.5)
}
\arguments{
\item{indice_masculinidad}{Indice de masculinidad del
conjunto de animales}

\item{umbral}{Umbral que define el indice de masculinidad}
}
\value{
Una lista indicando si el animal es freemartin de
acuerdo al umbral de masculinidad.
}
\description{
La función rectifica si un animal es o no un freemartin.
}
