% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_freem.R
\name{probabilidad_freemartin}
\alias{probabilidad_freemartin}
\title{Probabilidad de que un animal sea un freemartin}
\usage{
probabilidad_freemartin(hembras, machos, umbral = 0.5)
}
\arguments{
\item{hembras}{Numero de hembras}

\item{machos}{Numero de machos}

\item{umbral}{Umbral para considerar a un animal como un freemartin}
}
\value{
La probabilidad de que un animal sea un freemartin dependiendo
del umbral y del indice de masculinidad.
}
\description{
La función calcula la probabilidad de que un animal sea un
freemartin de acuerdo al umbral y al indice de masculinidad.
}
